Too early to abstract OAS3
